refactor winterm

+++ windows/install.ps1
@@ -19,6 +19,9 @@ Set-Location $BASEDIR
 # Run
 & $DOTBOT_BIN.Source -d "$BASEDIR\.." -c $CONFIG $args
 
-if ((Get-Content "$BASEDIR\apps\Microsoft.WindowsTerminal.Preview\settings.json").Count -gt (Get-Content "$HOME\AppData\Local\Packages\Microsoft.WindowsTerminalPreview_8wekyb3d8bbwe\LocalState\settings.json").Count) {
-    Copy-Item "$BASEDIR\apps\Microsoft.WindowsTerminal.Preview\settings.json" "$HOME\AppData\Local\Packages\Microsoft.WindowsTerminalPreview_8wekyb3d8bbwe\LocalState\settings.json"
+# Apps
+if (Test-Path "$HOME\AppData\Local\Packages\Microsoft.WindowsTerminalPreview_8wekyb3d8bbwe") {
+    if ((Get-Content "$BASEDIR\apps\Microsoft.WindowsTerminal.Preview\settings.json").Count -gt (Get-Content "$HOME\AppData\Local\Packages\Microsoft.WindowsTerminalPreview_8wekyb3d8bbwe\LocalState\settings.json").Count) {
+        Copy-Item "$BASEDIR\apps\Microsoft.WindowsTerminal.Preview\settings.json" "$HOME\AppData\Local\Packages\Microsoft.WindowsTerminalPreview_8wekyb3d8bbwe\LocalState\settings.json"
+    }
 }
